+++ packages/mongodb-chatbot-ui/package.json
@@ -81,7 +81,6 @@
     "@leafygreen-ui/icon": "^12.4.0",
     "@leafygreen-ui/icon-button": "^15.0.21",
     "@leafygreen-ui/leafygreen-provider": "^3.1.12",
-    "@leafygreen-ui/loading-indicator": "^2.0.12",
     "@leafygreen-ui/menu": "^22.0.9",
     "@leafygreen-ui/modal": "^16.0.7",
     "@leafygreen-ui/palette": "^4.0.10",
